update: Add config.go and put the configuration parsing in this file.
